+++ plugins/additional-severities.js
@@ -1,0 +1,17 @@
+
+function plugin(loggin) {
+    const { Severity } = loggin;
+
+    Severity
+        .register(0, 'EMERGENCY')
+        .register(1, 'ALERT')
+        .register(2, 'CRITICAL')
+        .register(3, 'ERROR')
+        .register(4, 'WARNING')
+        .register(5, 'NOTICE')
+        .register(6, 'INFO')
+        .register(7, 'DEBUG')
+        .register(8, 'SILLY');
+};
+
+module.exports = plugin;